Added cast to shut up compiler warning.

+++ src/lib/map.c
@@ -501,8 +501,8 @@ timeit(
 	elapsed = tm_elapsed_f(&end, &start);
 
 	if (verbose)
-		g_message("%s (%lu items, %u loop%s): %lf s (average: %lf s)", what,
-			(unsigned long) count, iter, iter == 1 ? "" : "s",
+		g_message("%s (%lu items, %lu loop%s): %lf s (average: %lf s)", what,
+			(unsigned long) count, (gulong) iter, iter == 1 ? "" : "s",
 			elapsed, elapsed / iter);
 
 	return elapsed;
